requirements.txt
netmiko
requests
pytest
xmltodict
pyaml